+++ backend/user/serializer.py
@@ -1,4 +1,5 @@
 from rest_framework import serializers
+from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
 
 from .models import User
 
@@ -6,3 +7,19 @@ class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ["username", "first_name", "last_name", "email", "institution", "experience", "approved"]
+
+class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
+    @classmethod
+    def get_token(cls, user):
+        token = super().get_token(user)
+
+        # Add custom claims
+        token['username'] = user.username
+        token['first_name'] = user.first_name      
+        token['last_name'] = user.last_name
+        token['email'] = user.email
+        token['institution'] = user.institution
+        token['experience'] = user.experience
+        token['approved'] = user.approved
+
+        return token
